feat: adicionando interfaces

+++ src/br/edu/ifsp/xyz/leitor/classe/CalcularComissaoTotal.java
@@ -1,12 +1,12 @@
 package br.edu.ifsp.xyz.leitor.classe;
 
 public class CalcularComissaoTotal {
-    private static Double comissaoTotal;
+    private static double comissaoTotal;
     private static String extratos;
     public static String CalcularComissaoTotal(Vendedor vendedor) throws Exception {
 
         extratos = "";
-        comissaoTotal = (double) 0;
+        comissaoTotal = 0;
         for(int i = 1; i <= 6; i++) {
             Pedido pedido = new Pedido("src/Pedido.txt", 0, String.valueOf(i));
             if (pedido.getIdVendedor()==vendedor.getIdVendedor()) {
